+++ src/Encoder/Exception/ExceptionInterface.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Rych\OTP\Encoder\Exception;
+
+interface ExceptionInterface extends \Throwable
+{
+}