+++ src/Encoder/Exception/RuntimeException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Rych\OTP\Encoder\Exception;
+
+class RuntimeException extends \RuntimeException implements ExceptionInterface
+{
+}